feat(Game): implement MVP MVVM architecture for book and inventory
- Added inventory events (InventoryUpdateEvent, InventoryItemAddedEvent, InventoryItemRemovedEvent) for reactive data updates.
- Added LocalPlayerReadyEvent to safely bind the UI to the local player's inventory upon gameplay start.
- Added ToggleBookEvent to manage the book's open/close state.
- Added InventoryUIRowType and a lightweight DTO structure (SUIItemData) for passing flat data.
- Created IBookViewListener and IInventoryPageViewListener interfaces for strict isolation between View and ViewModel layers.
- Implemented BookViewModel and InventoryPageViewModel. Added local caching and basic inventory sorting logic.
- Implemented BookUIView and InventoryPageUIView. Configured direct transfer of flat arrays into flash memory.
- Bound the "I" key press to trigger the book toggle.

+++ Game/Code/Components/Interaction/InteractableComponent.cpp
@@ -41,7 +41,7 @@ void CInteractableComponent::SetInteractionStringKey(const char* key)
 	m_interactionStringKey = StringUtils::GenerateID(key);
 }
 
-uint64 CInteractableComponent::GetInteractionStringKey() const
+uint32 CInteractableComponent::GetInteractionStringKey() const
 {
 	return m_interactionStringKey;
 }
